+++ src/builderUtils/resolvers.ts
@@ -1,18 +1,18 @@
-import type { OpenAPIV3 } from 'openapi-types';
+import type { OpenAPIV3_1 } from 'openapi-types';
 import { $ref2TypeName, isRefObject } from './converters';
 
 export const resolveParamsRef = (
-  openapi: OpenAPIV3.Document,
+  openapi: OpenAPIV3_1.Document,
   ref: string
-): OpenAPIV3.ParameterObject => {
+): OpenAPIV3_1.ParameterObject => {
   const target = openapi.components!.parameters![$ref2TypeName(ref).typeName];
   return isRefObject(target) ? resolveParamsRef(openapi, target.$ref) : target;
 };
 
 export const resolveSchemasRef = (
-  openapi: OpenAPIV3.Document,
+  openapi: OpenAPIV3_1.Document,
   ref: string
-): OpenAPIV3.SchemaObject => {
+): OpenAPIV3_1.SchemaObject => {
   const { typeName, propName } = $ref2TypeName(ref);
   let target = openapi.components!.schemas![typeName];
   target = !isRefObject(target) && propName ? target.properties![propName] : target;
@@ -20,17 +20,17 @@ export const resolveSchemasRef = (
 };
 
 export const resolveResRef = (
-  openapi: OpenAPIV3.Document,
+  openapi: OpenAPIV3_1.Document,
   ref: string
-): OpenAPIV3.ResponseObject => {
+): OpenAPIV3_1.ResponseObject => {
   const target = openapi.components!.responses![$ref2TypeName(ref).typeName];
   return isRefObject(target) ? resolveResRef(openapi, target.$ref) : target;
 };
 
 export const resolveReqRef = (
-  openapi: OpenAPIV3.Document,
+  openapi: OpenAPIV3_1.Document,
   ref: string
-): OpenAPIV3.RequestBodyObject => {
+): OpenAPIV3_1.RequestBodyObject => {
   const target = openapi.components!.requestBodies![$ref2TypeName(ref).typeName];
   return isRefObject(target) ? resolveReqRef(openapi, target.$ref) : target;
 };
